Polling interface on Submit trait.
The Submit traits two methods now have a poll-based interface, which
allows them to fail to prepare or submit an event, applying
backpressure when necessary.

The example driver, which is very bad in many ways, does not actually
apply backpressure right now.

This also involves a complete rewrite of the Submission future's
internals, to make it more explicitly a state machine. It also means
that the submission is *not* prepared until the future starts polling.
This is more consistent with the behavior of Rust's futures in general,
but it means that there's no way for individual events to indicate that
they should not be eagerly submitted. That will need to change later.