Add Ring interface.
The `Ring` interface wraps a normal IO object and runs its io on an
io-uring driver. It is intended to be the simplest high level API for
using io-uring.

=== The API ===

The Ring type takes an IO object and a driver to create a new async IO
object. The new object implements both AsyncRead and AsyncBufRead. A
simple example is demonstrated in `tests/ring-read.rs`

=== Current limitations ===

In this commit, the Ring type only supports Read operations. There are
several reasons for this:

* The buffered interface is the most efficient interface for IO
  operations on io-uring, because the Ring object is responsible for its
  own buffers. But there currently is no standard AsyncBufWrite trait.
* Cancelling an unfinished read and then starting a write (or vice
  versa) would cause Ring to panic right now. This API flaw needs to be
  resolved. The problem is explained below.

A read and write cannot be performed on the same buffer without
sequencing. This means if either a read or a write have already been
started on the buffer but were cancelled, starting the opposite
operation would be unacceptable. Instead, we need a better strategy for
handling these cancellation on persistent objects like this.
Investigation is ongoing.

We also don't have implemented the ability to perform opening and
closing operations on the fd using io-uring. These would need to be added
in the future.

=== Buffer management ===

The Ring type currently manages buffers by implementing a direct copy of
the BufReader type from std and futures, meaning it owns a buffer which
it uses again and again. This is far from the best design, and different
designs would be well-suited to different use cases. We should support
different buffer management strategies on the ring, introducing a third
parameter to the type. This needs further investigation as well.