New demo driver which give backpressure.
A new demo driver, which is now accessed through the function
`demo::driver()` instead of `&DRIVER`. The new driver is still the same
design (a blocking background thread processes all completions,
submissions are eagerly submitted with a syscall unless told not to),
which is probably not very performant, but it now implements
backpressure to prevent the user from preparing too many submissions at
once.

It does this using a new synchronization primitive built on top of a
WakerSet. I copied the slab-based WakerSet from async-std, but the
intrusive linked list approach used by tokio would be equally valid. I
used the slab-based approach because a) its API is safe and b) I think
using a slab is better than storing extra words of data in the future
states, for this use case.

The new AccessQueue primitive works by counting how many accesses to the
underlying data are "in flight," and only giving access if the number is
less than or equal to the amount that have been "released" to it. By
default, the API (which I imagine could be pulled into its own crate)
treats an ongoing reference, through the AccessGuard handle, as an
access, and uses an RAII pattern to manipulate the access count. We
don't use it that way, however: instead, we "hold" the access
"indefinitely," (usinga method so-named) to avoid dropping the access
immediately when we are done holding it. Then, when the completion
actually occurs, we release additional accesses.

In effect, no more IO events can be in flight than there is space to
complete them all on the submission queue. If there are too many in
flight, we will back off our preparation.