Fix Drive Pin guarantee soundness bug
The Drive trait intentionally pin projects, so that it can take
advantage of the pinning guarantee in its implementation. For example, a
valid Drive implementation could make use of an intrusive collection in
its definition.

However, I forgot about this and violated it by passing ownership of the
drive type out of Submission. This was done because when constructing,
for example, a File type, I need a driver for the file to run on.

Instead, file and tcpstream should use the same approach as tcplistener:
clone the drive handle used to set up the fd. This does not violate the
pin guarantees.