Cancel events on Ring before running new ones.
With this commit, ring now tracks when an event was cancelled, instead
of returning immediately to the Inert state. When preparing an event
from the Cancelled state, a cancel of the previous event is hard linked
before the first event being prepared.

This is intended to guarantee that all events scheduled using a single
Ring instance will be serialized (that is, they will occur
sequentially).