allow files to be opened and created with an extra set of flags
When dealing with fast storage systems, it is not unheard of to want to
open files with O_DIRECT, O_DSYNC, or a combination thereof.

This patch extends the existing interface to allow such flags to be
passed to open and create.

I have considered, instead of allowing for generic flags, to just add
another "nocache" variant, which would communicate intent better.

However usually systems have corner cases in which small adjustments
to the flag set must be done to accomodate for file system support,
etc. So I have decided for the flags version instead.

Please let me know if this is acceptable.

To use it:

        let mut output = File::create_with_extra_flags(filename, libc::O_DIRECT)
		.await
		.unwrap();