Add Cancel trait.
This provides a cleaner way to construct a Cancellation: it can now be
constructed from any type that implements Cancel, which includes things
like Boxes, CStrings, buffers, etc. All constructions of Cancellation
now pass through `Cancellation::from`, instead of unique constructors
for different types.

The Cancel trait provides a way to decompose an object into two words,
and then a way to drop that object from those two words. A Cancellation
is essentially `Box<dyn Drop>`, except it stores an extra word for
metadata for slices and trait objects.

Event::cancel has been changed to take ownership of the receiver, be
safe, and have a default implementation (which just leaks the receiver).
By default, therefore, the data owned by an event is leaked. An
implementation can instead construct a cancellation to deallocate the
data it owns.