fs: correctly handle File<D>::close()
We need to be very carefully once a IORING_OP_CLOSE request has been
submitted. If the IORING_OP_CLOSE has been completed when handling the
next request, it just panics! If the IORING_OP_CLOSE request is still
inflight, an IORING_OP_ASYNC_CANCEL is chained before the next request.
But the outcome of the request chain is undetermined, the file
descriptor may or may not have been closed. So panicking once the
IORING_OP_CLOSE request has been submitted, no matter is inflight or
completed. Also handle a corner case in cancellation.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>